Move apiritif sample extractor from nose plugin to Taurus
